refactor: extract oia config builders into config-transfer-builder service
